18-1-24

+++ Programs/Non_Uploaded/B_Erase_First_or_Second_Letter.cpp
@@ -33,11 +33,11 @@ using namespace std;
 
 int main() 
 {
-   int t ;
+   ll t ;
    cin >> t ;
    while(t--)
    {
-       int n ;
+       ll n ;
        cin >> n ;
        string s ;
        cin >> s ;
